requirements.txt
################################################################################
################################################################################
###                                                                          ###
###                       requirements.txt                                  ###
###                     Python Dependencies                                 ###
###                                                                          ###
################################################################################
################################################################################

Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
pyswisseph==2.10.3.2
googlemaps==4.10.0
pytz==2024.1
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
requests==2.31.0
cryptography==42.0.5
